+++ src/models/AuditJavascript.php
@@ -15,6 +15,7 @@ use Yii;
  * @property int    $id
  * @property int    $entry_id
  * @property string $created
+ * @property string $type
  * @property string $message
  * @property string $origin
  * @property string $data
@@ -29,7 +30,7 @@ class AuditJavascript extends ActiveRecord
      */
     public static function tableName()
     {
-        return '{{audit_javascript}}';
+        return '{{%audit_javascript}}';
     }
 
     /**
@@ -49,6 +50,7 @@ class AuditJavascript extends ActiveRecord
             'id'        => Yii::t('audit', 'ID'),
             'entry_id'  => Yii::t('audit', 'Entry ID'),
             'created'   => Yii::t('audit', 'Created'),
+            'type'      => Yii::t('audit', 'Type'),
             'message'   => Yii::t('audit', 'Message'),
             'origin'    => Yii::t('audit', 'Origin'),
             'data'      => Yii::t('audit', 'Data'),